requirements.txt
Flask>=1.1.0,<2.0
Flask-SQLAlchemy>=2.4.0
Flask-RESTful>=0.3.8
Flask-Marshmallow>=0.14.0
marshmallow-sqlalchemy>=0.23.0
Flask-JWT-Extended>=3.24.0
Werkzeug>=1.0.0,<2.0
psycopg2-binary>=2.8.0
gunicorn>=20.0.0
python-dotenv>=0.15.0
email-validator>=1.1.0